feat: add directors feature

+++ src/main/java/ru/yandex/practicum/filmorate/Service/DirectorImpl.java
@@ -5,10 +5,7 @@ import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.Repository.DirectorRepository;
 import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Director;
-
 import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -32,11 +29,17 @@ public class DirectorImpl implements DirectorService{
 
     @Override
     public Director update(Director director) {
+        if(directorRepository.findById(director.getId()).isEmpty()) {
+            throw new NotFoundException("No director with id: " + director.getId());
+        }
         return directorRepository.update(director);
     }
 
     @Override
     public boolean delete(int directorId) {
+        if(directorRepository.findById(directorId).isEmpty()) {
+            throw new NotFoundException("No director with id: " + directorId);
+        }
         return directorRepository.delete(directorId);
     }
 }
